refactor: disambiguate session names
Rename src/lib/sessions.functions.ts to chatSessions.functions.ts and
src/server/session.ts to sessionCookie.ts so the ADK chat sessions and
the BFF auth session cookie no longer share an ambiguous "session" name.
