+++ src/lib/chatSessions.functions.ts
